+++ libs/ui/src/lib/entity-logo/entity-logo.component.ts
@@ -1,3 +1,4 @@
+import { getEmojiFlag } from '@ghostfolio/common/helper';
 import { EntityLogoImageSourceService } from '@ghostfolio/ui/entity-logo/entity-logo-image-source.service';
 
 import {
@@ -11,12 +12,14 @@ import { DataSource } from '@prisma/client';
 
 @Component({
   changeDetection: ChangeDetectionStrategy.OnPush,
+  host: { class: 'align-items-center d-flex' },
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
   selector: 'gf-entity-logo',
   styleUrls: ['./entity-logo.component.scss'],
   templateUrl: './entity-logo.component.html'
 })
 export class GfEntityLogoComponent implements OnChanges {
+  @Input() countryCode: string;
   @Input() dataSource: DataSource;
   @Input() hasPlaceholder = false;
   @Input() size: 'large';
@@ -24,6 +27,7 @@ export class GfEntityLogoComponent implements OnChanges {
   @Input() tooltip: string;
   @Input() url: string;
 
+  public emojiFlag?: string;
   public hasError = false;
   public src?: string;
 
@@ -32,17 +36,19 @@ export class GfEntityLogoComponent implements OnChanges {
   ) {}
 
   public ngOnChanges() {
+    this.emojiFlag = undefined;
     this.hasError = false;
+    this.src = undefined;
 
-    if (this.dataSource && this.symbol) {
+    if (this.countryCode) {
+      this.emojiFlag = getEmojiFlag(this.countryCode);
+    } else if (this.dataSource && this.symbol) {
       this.src = this.imageSourceService.getLogoUrlByAssetProfileIdentifier({
         dataSource: this.dataSource,
         symbol: this.symbol
       });
     } else if (this.url) {
       this.src = this.imageSourceService.getLogoUrlByUrl(this.url);
-    } else {
-      this.src = undefined;
     }
   }